Change the behavior of opening the output file
Access to the output filehandle is extracted to the _output_fh() method
and saved in the 'output_fh' hash key.  prepare_app() now saves the output
file name in the 'output_filename' hash key.

Opening the output filehandle is now delayed until the first write.  If the
PSGI environment 'psgi.run_once' is true, then assume we're running as a CGI
script, and open the file in append mode.  This way, independent executions of
the CGI script will string all the sessions together into the same output file.

If psgi.run_once is false, then it opens the file in write mode, overwriting
whatever was there before.